Restart

+++ homework3/3-parallel/3-parallel-hashtable/src/gpu_hashtable.cu
@@ -209,7 +209,7 @@ bool GpuHashTable::insertBatch(int *keys, int* values, int numKeys) {
 
 	this->size += numKeys;
 
-	int newf = (this->size + numKeys) / this->hmax;
+	double newf = (this->size + numKeys) / this->hmax;
 	cout << "New hmax: " << this->hmax << endl;
 	cout << "Num keys: " << numKeys << endl;
 	cout << "New size: " << this->size << endl;
